INTYGFV-16702: Moved ActionRuleBlockTestIndicatedPerson into separate class. Added @Transactional to ForwardCertificateService.
