+++ certificate-service/domain/src/main/java/se/inera/intyg/certificateservice/domain/action/certificate/model/CertificateActionSign.java
@@ -7,7 +7,6 @@ import lombok.Builder;
 import lombok.Getter;
 import se.inera.intyg.certificateservice.domain.certificate.model.Certificate;
 import se.inera.intyg.certificateservice.domain.certificatemodel.model.CertificateActionSpecification;
-import se.inera.intyg.certificateservice.domain.patient.model.Patient;
 
 @Builder
 @Getter(AccessLevel.NONE)
@@ -67,41 +66,5 @@ public class CertificateActionSign implements CertificateAction {
             }
         )
         .orElse(SIGN_DESCRIPTION);
-  }
-
-  public static class ActionRuleBlockTestIndicatedPerson implements ActionRule {
-
-    @Override
-    public boolean evaluate(Optional<Certificate> certificate,
-        Optional<ActionEvaluation> actionEvaluation) {
-      if (certificate.isEmpty()) {
-        return evaluate(actionEvaluation);
-      }
-
-      return actionEvaluation.filter(
-          evaluation ->
-              certificate.filter(
-                      value -> isPatientNotTestIndicated(value.certificateMetaData().patient())
-                  )
-                  .isPresent()
-      ).isPresent();
-    }
-
-    @Override
-    public String getReasonForPermissionDenied() {
-      return "Åtgärden kan inte utföras för testpersoner.";
-    }
-
-    private static boolean evaluate(Optional<ActionEvaluation> actionEvaluation) {
-      return actionEvaluation.filter(
-              evaluation -> isPatientNotTestIndicated(evaluation.patient())
-          )
-          .isPresent();
-    }
-
-    private static boolean isPatientNotTestIndicated(Patient patient) {
-      return !patient.testIndicated().value();
-    }
-
   }
 }
